I made movement script
The movement script uses the new Unity input system so I added that package to the project. I also added a scripts folder and a player folder within it. I then made two objects for a player in floor. Then I made a player controls player input manager. This generated a script to go with it. I then made a jump fix that is a little broken. Its meant to allow for a short and tall jump. It increases gravity during falls to make jumps feel more impactful. The movement script only adds movement along x-axis as well as the jump. My thought is other movement that's character specific can be done through inheriting the main player movement script and then using a separate script.